requirements.txt
########  CORE  ########
Django>=4.2,<5.0
gunicorn>=20.1
psycopg2-binary>=2.9
django-environ>=0.11
whitenoise>=6.6

########  ASYNC / QUEUE  ########
celery[redis]>=5.3
redis>=5.0
channels>=4.0
django-redis>=5.4
django-celery-beat

########  REST / AUTH  ########
djangorestframework>=3.15
django-cors-headers>=4.3
django-widget-tweaks>=1.5
python-dotenv>=1.0

########  HTTP & INTEGRACIJE  ########
requests>=2.31
urllib3>=2.2
boto3>=1.34       # S3 back-up ili MinIO

########  DATE / TIME / TZ  ########
python-dateutil>=2.9
pytz>=2024.1

########  EXCEL / CSV / DATA  ########
pandas>=2.2
numpy>=1.26
openpyxl>=3.1
xlsxwriter>=3.2
xlrd>=2.0
xlwt>=1.3

########  PDF / IZVJEŠTAJI  ########
reportlab>=4.1
weasyprint>=60.2
Pillow>=10.3

########  GRAFI / ANALITIKA (scoreboard)  ########
matplotlib>=3.9
